ukb_analysis_template/requirements.txt
pandas>=2.0
numpy>=1.24
pyyaml>=6.0
statsmodels>=0.14
lifelines>=0.28
scikit-learn>=1.3
pytest>=8.0
